Added support for named arguments and default arguments in applyRccClosure.
The argument matching code in the previous version was commented out, presumable for efficiency reasons. (Handling default/named arguments is an extra few passes over the argument list.) I moved the version without argument matching to a new function 'applyPlainRccClosure', which is currently unused but will come in handy for compile-time argument matching.


git-svn-id: file:///home/john/Code/repo-dumps/R-2.1.1rcc/trunk@11 b3340ace-0d48-454c-a616-96b99b0299b0