Fixed bug in symbol table lookup, which was failing for uses of internal R names.
Removed SymbolTableAnnotationMap. SymbolTable is now a field of LexicalScope. (We need a symbol table for each lexical scope, not just the procedures. This has the added benefit of being easier to use.)

SymbolTables are now populated when VarBindingAnnotationMap is computed.

"first name" field of a FuncInfo is now a CONS cell.

Added VarBindings for formal arguments as well as mentions.

Added VarAnnotationMap::is_valid().

Now produced call graph debugging output in DOT form.


git-svn-id: file:///home/john/Code/repo-dumps/rcc/trunk@250 925864f1-7c04-44ca-bbaa-d8d9a55e96f5