Removed other people's code from regression test directory.
While it's certainly important to check RCC for regressions against real-world R programs, we shouldn't publicize RCC with other people's code in the test directory.

If a regression is found in real-world code and not in the regression directory, we should write a simple regression test in the tests/regression directory that exhibits the bug.



git-svn-id: file:///home/john/Code/repo-dumps/rcc/trunk@307 925864f1-7c04-44ca-bbaa-d8d9a55e96f5